Create `useRedemptionEstimatedFees` hook
Creates `useRedemptionEstimatedFees` hook where we can get the estimated BTC
amount that we will receive from a redemption and threshold network fee that we
will pay during a redemption.
