+++ src/hooks/tbtc/index.ts
@@ -1,10 +1,11 @@
-export * from "./useRevealDepositTransaction"
-export * from "./useTBTCDepositDataFromLocalStorage"
-export * from "./useTBTCVaultContract"
-export * from "./useSubscribeToOptimisticMintingFinalizedEvent"
-export * from "./useSubsribeToDepositRevealedEvent"
-export * from "./useSubscribeToOptimisticMintingRequestedEvent"
 export * from "./useFetchDepositDetails"
 export * from "./useFetchRecentDeposits"
 export * from "./useFetchTBTCMetrics"
+export * from "./useRedemptionEstimatedFees"
 export * from "./useRequestRedemption"
+export * from "./useRevealDepositTransaction"
+export * from "./useSubscribeToOptimisticMintingFinalizedEvent"
+export * from "./useSubscribeToOptimisticMintingRequestedEvent"
+export * from "./useSubsribeToDepositRevealedEvent"
+export * from "./useTBTCDepositDataFromLocalStorage"
+export * from "./useTBTCVaultContract"
